Export dumpers to a separate file. Add KML and CSV dumpers.
