+++ igc_lib_demo.py
@@ -3,6 +3,7 @@ import os
 import sys
 
 import igc_lib
+import dumpers
 
 
 if __name__ == "__main__":
@@ -14,6 +15,9 @@ if __name__ == "__main__":
     input_base_file = os.path.splitext(input_file)[0]
     wpt_file = "%s-thermals.wpt" % input_base_file
     cup_file = "%s-thermals.cup" % input_base_file
+    thermals_csv_file = "%s-thermals.csv" % input_base_file
+    flight_csv_file = "%s-flight.csv" % input_base_file
+    kml_file = "%s-flight.kml" % input_base_file
 
     flight = igc_lib.Flight.create_from_file(input_file)
     print "flight =", flight
@@ -22,6 +26,10 @@ if __name__ == "__main__":
         print "glide[%d] " % x, glide
         print "thermals[%d] = " % x, thermal
 
-    print "Dumping thermals to %s and %s" % (wpt_file, cup_file)
-    flight.dump_thermals_to_wpt_file(wpt_file, True)
-    flight.dump_thermals_to_cup_file(cup_file)
+    print "Dumping thermals to %s, %s and %s" % (wpt_file, cup_file, thermals_csv_file)
+    dumpers.dump_thermals_to_wpt_file(flight, wpt_file, True)
+    dumpers.dump_thermals_to_cup_file(flight, cup_file)
+
+    print "Dumping flight to %s and %s" % (kml_file, flight_csv_file)
+    dumpers.dump_flight_to_csv(flight, flight_csv_file, thermals_csv_file)
+    dumpers.dump_flight_to_kml(flight, kml_file)
